requirements.txt
# Core
gymnasium==0.29.1
numpy>=1.24.0

# LLM
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
pydantic>=2.0.0

# OpenTelemetry (Single Source of Truth)
# All observability (including API metrics) via OpenTelemetry to Langfuse/LangSmith
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-instrumentation>=0.42b0
opentelemetry-instrumentation-langchain>=0.1.0  # Optional: LangChain auto-instrumentation

# Visualization
streamlit>=1.31.0
plotly>=5.18.0
pandas>=2.0.0

# Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
packaging>=21.0
tqdm>=4.65.0

# Database
# SQLite3 is built-in to Python, no external dependency needed

# GraphRAG (Langfuse GraphRAG side project)
graphrag>=0.1.0  # Microsoft GraphRAG library
neo4j>=5.11.0  # Neo4j Python driver with vector support